refactor: move packer/unpacker to internal
Removes buildcfg dependency from pkg directory
