+++ internal/pkg/image/unpacker/squashfs_test.go
